requirements.txt
praw==7.7.1
requests
pandas
numpy
sentence-transformers
chromadb
ruptures
scipy
langchain
langchain-community
langgraph
streamlit
plotly
python-dotenv
transformers
groq
